requirements.txt
streamlit==1.41.1
pandas==2.2.3
python-dateutil==2.9.0.post0